+++ CHANGELOG.rst
@@ -39,6 +39,7 @@ Changed
 - feat: added `to_json` method for `DatetimeSinusoidCalculator` `#556 <https://github.com/azukds/tubular/issues/556>`_
 - feat: added a new transformer called DatetimeComponentExtractor
 - feat: add a decorator to register transformer classes by name `#588 <https://github.com/azukds/tubular/issues/588>_`
+- feat: add utility class for serializing and deserializing pipelines `#588 <https://github.com/azukds/tubular/issues/588>_`